fix:修正匯出某筆的內部料號及後面的資料都沒有的情況
fix:修正AI執行排程時，錯誤不會寫進CycleErrorLogDetail資料表裡面的問題

+++ SanderQuotation/ViewModel/QuotationResult/QuotationItemVM.cs
@@ -18,8 +18,8 @@ namespace ViewModel.QuotationResult
         /// <summary>元件描述</summary>
         public string? Description { get; set; }
 
-        /// <summary>BOM 數量</summary>
-        public int? Qty { get; set; }
+        /// <summary>BOM 數量（允許小數）</summary>
+        public decimal? Qty { get; set; }
 
         /// <summary>廠商</summary>
         public string? Manufacturer { get; set; }
